--- packages/components/modal/src/Modal.styles.ts
@@ -1,8 +1,0 @@
-import { css } from 'emotion';
-import tokens from '@contentful/f36-tokens';
-
-export const styles = {
-  modal: css({
-    color: tokens.colorRedBase,
-  }),
-};
