+++ packages/components/modal/src/index.ts
@@ -1,2 +1,11 @@
 export { Modal } from './Modal';
 export type { ModalProps } from './Modal';
+export { ModalConfirm } from './ModalConfirm/ModalConfirm';
+export type { ModalConfirmProps } from './ModalConfirm/ModalConfirm';
+export { ModalContent } from './ModalContent/ModalContent';
+export type { ModalContentProps } from './ModalContent/ModalContent';
+export { ModalControls } from './ModalControls/ModalControls';
+export type { ModalControlsProps } from './ModalControls/ModalControls';
+export { ModalHeader } from './ModalHeader/ModalHeader';
+export type { ModalHeaderProps } from './ModalHeader/ModalHeader';
+export { ModalLauncher } from './ModalLauncher/ModalLauncher';
